app/shared/pyproject.toml
[tool.poetry]
name = "shared"
version = "0.2.4"
description = "Carrot shared library"
authors = ["Sam Cox <sam.cox@nottingham.ac.uk>", "Philip Quinlan <philip.quinlan@nottingham.ac.uk>"]
license = "MIT"
packages = [
  {include = "shared"},
  {include = "shared/py.typed"},
]

[tool.poetry.dependencies]
python = "^3.11"
django = "~4.2.0"
psycopg2-binary = "^2.9.9"
azure-storage-queue = "^12.9.0"
python-dotenv = "^1.0.1"
graphviz = "^0.20.3"
azure-storage-blob = "12.20.0"
djangorestframework = "3.15.2"
django-filter = "24.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
psycopg2 = "^2.9.9"
django-stubs = "^5.0.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = ["."]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
